+++ lib/app/modules/mobcol/dashboard/binding/mobile_collection_binding.dart
@@ -1,5 +1,5 @@
 import 'package:get/get.dart';
-import 'package:sufi_one/app/modules/mobcol/controllers/mobile_collection_controller.dart';
+import 'package:sufi_one/app/modules/mobcol/dashboard/controllers/mobile_collection_controller.dart';
 
 class MobileCollectionBinding extends Bindings {
   @override
