+++ lib/app/modules/mobcol/tugasBaru/binding/mobcol_kunjungan_binding.dart
@@ -1,5 +1,5 @@
 import 'package:get/get.dart';
-import 'package:sufi_one/app/modules/mobcol/controllers/mobcol_kunjungan_controller.dart';
+import 'package:sufi_one/app/modules/mobcol/tugasBaru/controller/mobcol_kunjungan_controller.dart';
 
 
 class KunjunganBinding extends Bindings {
