+++ lib/app/modules/mobcol/tugasBaru/binding/mobcol_tugasbaru_binding.dart
@@ -1,5 +1,5 @@
 import 'package:get/get.dart';
-import 'package:sufi_one/app/modules/mobcol/controllers/mobcol_tugasbaru_controller.dart';
+import 'package:sufi_one/app/modules/mobcol/tugasBaru/controller/mobcol_tugasbaru_controller.dart';
 
 
 class MobcolTugasbaruBinding extends Bindings {
