+++ lib/app/modules/mobcol/tugasBelumSelesai/binding/mobcol_tugasbelumselesai_bindings.dart
@@ -1,5 +1,5 @@
 import 'package:get/get.dart';
-import 'package:sufi_one/app/modules/mobcol/controllers/mobcol_tugasbelumselesai_controller.dart';
+import 'package:sufi_one/app/modules/mobcol/tugasBelumSelesai/controller/mobcol_tugasbelumselesai_controller.dart';
 
 
 class MobcolTugasbelumselesaiBinding extends Bindings {
